fix(linux): R20-tail — portable compile fixes surfaced by the engine-less Linux build
Windows compiled these only because the engine/vendor trees were present:
- Capability.cpp: the #else (no HAS_RAPIDOCR) branch of probeOcrRapidModels
  used `c` without declaring it; declaration hoisted above the guard.
- PageLabels.cpp + TestPageLabels.cpp + TestLinkBookmarkRoundTrip.cpp:
  PdfObject(static_cast<long long>) is ambiguous on LP64 (int64_t is long
  there, and long long also converts to PdfObject(double));
  static_cast<std::int64_t> matches the exact overload on both ABIs.
- BackendRouter.cpp: destroying unique_ptr<IPdfRenderer> needs the complete
  type in EVERY configuration; it arrived only via PdfiumBackend.h under
  HAS_PDFIUM. Include core/interfaces/IPdfRenderer.h directly.

+++ src/core/PageLabels.cpp
@@ -133,7 +133,11 @@ bool writeNumberTree(PoDoFo::PdfMemDocument& doc, int startValue, Style style,
         auto& labels = objects.CreateDictionaryObject();
         PoDoFo::PdfArray nums;
         for (const PageLabelNumEntry& e : entries) {
-            nums.Add(PoDoFo::PdfObject(static_cast<long long>(e.pageNum)));
+            // std::int64_t, not `long long`: PoDoFo 1.1's PdfObject has an
+            // exact int64_t overload, and on LP64 (Linux) int64_t is `long`,
+            // so a `long long` argument was ambiguous against PdfObject(double)
+            // (a Windows-lane compile success that broke the Linux build).
+            nums.Add(PoDoFo::PdfObject(static_cast<std::int64_t>(e.pageNum)));
 
             PoDoFo::PdfObject range{PoDoFo::PdfDictionary()};
             range.GetDictionary().AddKey(
@@ -141,7 +145,7 @@ bool writeNumberTree(PoDoFo::PdfMemDocument& doc, int startValue, Style style,
             // /St is spec-defaulted to 1, but writing it explicitly keeps
             // the readback exact (no default-reconstruction in consumers).
             range.GetDictionary().AddKey(
-                "St", PoDoFo::PdfObject(static_cast<long long>(e.startValue)));
+                "St", PoDoFo::PdfObject(static_cast<std::int64_t>(e.startValue)));
             nums.Add(range);
         }
         labels.GetDictionary().AddKey("Nums", PoDoFo::PdfObject(nums));
